% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_key_stats.R
\name{create_key_stats}
\alias{create_key_stats}
\title{Create key stats}
\usage{
create_key_stats()
}
\arguments{
\item{none}{none}
}
\description{
Creates statistics for collected keys
}
\examples{
create_key_stats()
}
\keyword{Fide}
\keyword{list}
\keyword{players}
\keyword{statistics}

